+++ packages/jala_core/lib/jala_core.dart
@@ -39,3 +39,4 @@ export 'src/throttle/jala_throttle_profile.dart';
 export 'src/throttle/jala_throttle_registry.dart';
 export 'src/util/glob.dart';
 export 'src/util/id_generator.dart';
+export 'src/util/query_params.dart';